requirements.txt
# Основные зависимости для тестирования
pytest==7.4.3
requests==2.32.4

# Для работы с Binance Testnet
python-binance==1.0.19
python-dotenv==1.0.0

# Отчеты и документация
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.5.0  # для параллельного запуска тестов

# Моки и тестирование
pytest-mock==3.11.1
responses==0.24.1  # для мокинга HTTP-запросов
pip~=25.3
typing_extensions~=4.12.2
attrs~=24.2.0
protobuf~=6.31.1
cryptography~=44.0.1
docutils~=0.21.2
pytz~=2025.2
PyYAML~=6.0.2
idna~=3.8
multidict~=6.0.5
regex~=2025.11.3
execnet~=2.1.2
setproctitle~=1.3.3
pycryptodome~=3.23.0
cffi~=1.17.0
click~=8.1.7
MarkupSafe~=2.1.5
pluggy~=1.5.0
Deprecated~=1.2.14
iniconfig~=2.3.0
importlib_metadata~=8.0.0
argcomplete~=3.5.0
yarl~=1.9.4
gunicorn~=23.0.0
aiohttp~=3.13.3
aiosignal~=1.3.1
frozenlist~=1.4.1
aiohappyeyeballs~=2.4.0
dateparser~=1.2.2
websockets~=16.0
certifi~=2024.7.4
tzlocal~=5.3.1
urllib3~=2.6.3
six~=1.16.0
python-dateutil~=2.9.0.post0
charset-normalizer~=3.3.2
allure-python-commons~=2.13.2
packaging~=24.1
Werkzeug~=2.2.3
